Progress tracking for quiz
Quiz now submits progress as well, adds a 150% (50% increase) multiplier
to the round score depending on the average score of all your quiz
attempts that round. Example, you get 50%, then 100%, you get 70%.
because kids also lose 5% per attempt

+++ Assets/Scripts/FrogScripts/FrogGF.cs
@@ -99,7 +99,7 @@ public class FrogGF : GameFlow {
         textPanel.SetText(t0);// + t1 + t2 + t3 + t4);
         textPanel.StartWriting();
 		scoreText.gameObject.SetActive (false);
-        float scorePerc = Mathf.Min(matureDescendants / frogsForMaxScore, 1);
+        float scorePerc = Mathf.Min((float)matureDescendants / (float)frogsForMaxScore, 1);
         ProgressTracker.Instance.SetRoundScore(scorePerc, 0);
         ProgressTracker.Instance.SubmitProgress(0);        
     }
